app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "jsonpatch>=1.33",
    "pywebview[qt5]>=6.0 ; sys_platform == 'linux'",
    "pywebview[qt]>=6.0 ; sys_platform != 'linux'",
    "sqlmodel>=0.0.24",
    "websockets>=15.0.1",
]
